+++ tcs_bin_path.pl
@@ -1,6 +1,0 @@
-#!/usr/bin/env perl
-use strict;
-use warnings;
-use FindBin;
-my $mypath = $FindBin::Bin;
-print( $mypath );