+++ tools/gen_autotools_acam.pl
@@ -8,7 +8,7 @@
 ## - If any files have been updated, it will exit with 0.
 ## - If no files have been updated, it will exit with a value other than 0.
 ##
-## - $Revision: 1.9 $
+## - $Revision: 1.10 $
 ##
 ## - Author: 2025-2026, tomyama
 ## - Intended primarily for personal use, but BSD license permits redistribution.